+++ Code/ConfigOptions.cs
@@ -11,6 +11,8 @@ namespace LordsItemEdits
         public static ConfigEntry<bool> EnableVoidDiosEdit;
         public static ConfigEntry<bool> AllowRespawnAsVoidReaver;
 
+        public static ConfigEntry<bool> EnableAtgEdit;
+
         internal static void BindConfigOptions(ConfigFile config)
         {
             EnableVoidDiosEdit = config.BindOption(
@@ -25,6 +27,15 @@ namespace LordsItemEdits
                 "Allow respawning as a Void Reaver",
                 "Should the Void Reaver be an option to respawn as with the edited Pluripotent Larva? This is configurable since their only attack has bad damage and can't target flying enemies.",
                 false
+            );
+
+
+            EnableAtgEdit = config.BindOption(
+                "ATG",
+                "Enable edit",
+                "Enable the ATG edit?",
+                true,
+                Extensions.ConfigFlags.RestartRequired
             );
         }
     }